modularize react example components
