+++ packages/react-client/examples/vite-example/index.html
@@ -2,7 +2,11 @@
 <html lang="en">
     <head>
         <meta charset="UTF-8" />
-        <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+        <link
+            rel="icon"
+            type="image/svg+xml"
+            href="https://www.hashport.network/wp-content/uploads/hashport-glyph-light-1.svg"
+        />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>hashport</title>
     </head>
