+++ packages/react-client/examples/vite-example/src/App.css
@@ -5,43 +5,21 @@
   text-align: center;
 }
 
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-  transition: filter 300ms;
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5em
 }
 
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
+.button-group {
+  display: flex;
+  justify-content: center;
+  gap: 0.5em;
 }
 
-.logo.react:hover {
-  filter: drop-shadow(0 0 2em #61dafbaa);
-}
-
-@keyframes logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  a:nth-of-type(2) .logo {
-    animation: logo-spin infinite 20s linear;
-  }
-}
-
-.card {
-  padding: 2em;
-}
-
-.read-the-docs {
-  color: #888;
+.status-text {
+  color: rgb(107, 129, 255);
+  font-style: italic;
 }
 
 h1,
@@ -51,10 +29,5 @@ h4,
 h5,
 h6,
 p {
-  margin-block: 0.5rem
-}
-
-h5 span {
-  color: rgb(190, 83, 190);
-  font-style: italic;
+  margin-block: 1rem
 }
