+++ packages/react-client/examples/vite-example/src/components/TransactionStatus.tsx
@@ -1,0 +1,11 @@
+import { useProcessingTransaction } from '@hashport/react-client';
+
+export const TransactionStatus = () => {
+    const processingTx = useProcessingTransaction();
+
+    return (
+        <p>
+            Transaction Status: <span className="status-text">{processingTx.status}</span>
+        </p>
+    );
+};
